src/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
firebase-admin==5.3.0
Pillow==11.1.0
torch==2.0.0
torchvision==0.15.0
opencv-python-headless==4.11.0.86
numpy==2.2.4
ultralytics==8.0.0
functions-framework==3.* 